Code360
Problem statement
You are given an array of strings “arr” and a character matrix “mat”. Your task is to find out all the possible words in the array “arr” that can be constructed by moving sequentially in the matrix “mat”, where the movement can start from any position in the matrix “mat” and can be done in all 8 possible directions. However, the same cell can not be considered more than once for the construction of a word. Print the words in sorted manner.

For Example:
If arr = {“go”, “pp”, “no”, “op”}, mat = {{‘g’,’p’}, 
                                          {‘d’,’o’}}.
Here, the two-letter words that can be formed from the matrix “mat” are “gp”, “gd”, “go”, “pg”, “pd”, “po”, “dg”, “dp”, “do”, “og”, “op”, “od”.

So, the words that are present in the array “arr” which can be formed from the matrix “mat” are “go” and “op”.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 10
1 <= P <= 10
1 <= arr[i].length <= 50
1 <= N, M <= 7
All the characters of the strings in the array “arr” and matrix “mat” contain lowercase English letters only.
All the words in the array “arr” are unique.

Time limit: 1 second
Sample Input 1:
1
6
go on no of pon pgp
2 2
g n
o p
Sample Output 1:
go no on pon
Explanation for sample 1:
For the first test case, the word “go” can be formed by (0,0) → (1,0).
The word “on” can be formed by (1,0) → (0,1).
The word “no” can be formed by (0,1) → (1,0).
The word “of” can never be formed as there is no ‘f’ character in the matrix.
The word “pon” can be formed by (1,1) → (1,0) → (0,1).
The word “pgp” can not be formed as there is only one ‘p’ character in the matrix.
Sample Input 2:
1
8
person so warm on code no son rn  
2 3
p e r
n o s
Sample Output 2:
no on person so son
Explanation for sample 2:
For the first test case, the word “person” can be formed by (0,0) → (0,1) → (0,2) → (1,2) → (1,1) → (1,0).
The word “so” can be formed by (1,2) → (1,1).
The word “warm” can never be formed as there is no character ‘w’ in the matrix.
The word “on” can be formed by (1,1) → (1,0).
The word “code” can never be formed as there is no character ‘c’ in the matrix.
The word “no” can be formed by (1,0) → (1,1).
The word “son” can be formed by (1,2) → (1,1) → (1,0).
The word “rn” can not be formed as the character ‘r’ and ‘n’ are not adjacent to each other.